python/smolagents/zero-bloat-jq-filter/pyproject.toml
[project]
name = "autonomous-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "litellm>=1.83.13",
    "mcp>=1.27.0",
    "python-dotenv>=1.0.1",
    "smolagents>=1.24.0",
]
